+++ packages/core/src/js/index.js
@@ -38,6 +38,7 @@ export { installFieldErrors } from './field-errors.js';
 export { installCsrfHeader } from './csrf-header.js';
 export { installCopy } from './copy.js';
 export { installSpy } from './spy.js';
+export { installNavCurrent } from './nav-current.js';
 export { installSparkline } from './sparkline.js';
 export { installCodeEditor } from './code-editor.js';